% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anatomic_stage.R
\name{anatomic_stage}
\alias{anatomic_stage}
\title{Calculate Anatomic Stage from TNM Measures on imaging/biopsy}
\usage{
anatomic_stage(t, n, m, t_mi = F, n_mi = F)
}
\arguments{
\item{t}{Clinical T (tumor) stage between 0 and 4}

\item{n}{Clinical N (nodal) stage between 0 and 3}

\item{m}{Clinical M (metastatic) stage, either TRUE or FALSE}

\item{t_mi}{Clinical microinvasive tumor, either TRUE or FALSE. Only ever
possibly TRUE when tumor stage is T1}

\item{n_mi}{Clinical micrometastatic nodal disease, either TRUE or FALSE.
Only ever possibly TRUE when nodal stage is N1}
}
\value{
a character string representing clinical prognostic stage based on
the above variables
}
\description{
Calculate Anatomic Stage from TNM Measures on imaging/biopsy
}
\examples{
anatomic_stage(1,0,F)
}
